fix(financial-facts): latest-period-end pick everywhere a fiscal period has comparative candidates
A filing re-reports comparative prior periods under its own fiscal
identity; picking by filed date alone surfaces a comparative column
(#1546). The MCP fact/series tools and the web Financials tab now share
the span-aware, latest-ending, latest-filed pick via
StatementLineFacts.PickCurrentlyReported.
