+++ src/Equibles.Sec.FinancialFacts.Data/Enums/ConceptBalance.cs
@@ -1,5 +1,3 @@
-using System.ComponentModel.DataAnnotations;
-
 namespace Equibles.Sec.FinancialFacts.Data.Enums;
 
 /// <summary>
