+++ src/Equibles.Sec.FinancialFacts.Mcp/Tools/FinancialStatementTools.cs
@@ -293,27 +293,7 @@ public class FinancialStatementTools
     internal static FinancialFact PickCurrentlyReportedFact(
         IEnumerable<FinancialFact> facts,
         SecFiscalPeriod fiscalPeriod
-    )
-    {
-        var candidates = facts.ToList();
-
-        var preferred = candidates
-            .Where(f =>
-            {
-                var spanDays = f.PeriodEnd.DayNumber - f.PeriodStart.DayNumber;
-                return fiscalPeriod == SecFiscalPeriod.FullYear
-                    ? spanDays == 0 || spanDays >= MinAnnualSpanDays
-                    : spanDays <= MaxDiscreteQuarterDays;
-            })
-            .ToList();
-        if (preferred.Count > 0)
-            candidates = preferred;
-
-        return candidates
-            .OrderByDescending(f => f.PeriodEnd)
-            .ThenByDescending(f => f.FiledDate)
-            .First();
-    }
+    ) => StatementLineFacts.PickCurrentlyReported(facts, fiscalPeriod);
 
     // Chronological order within a fiscal year: Q1 < Q2 < Q3 < Q4 < FullYear.
     private static int ChronologicalRank(SecFiscalPeriod period) =>
